pyproject.toml
[project]
name = "english-conversation-practice-app"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "eijun", email = "eijun.tei@gmail.com" }
]
dependencies = [
    "streamlit>=1.45.1",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.17",
    "pydub>=0.25.1",
    "speechrecognition>=3.14.3",
    "gtts>=2.5.4",
    "streamlit-webrtc>=0.62.4",
    "openai-whisper>=20240930",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.5",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/english_conversation_practice_app"]
